Logics for statics were added

+++ yoAprendo-api/src/main/java/com/joacocenteno/yoAprendo_api/dto/ProgressResponse.java
@@ -10,5 +10,5 @@ public class ProgressResponse {
     Long user_id;
     Long lesson_id;
     PorgressStatus status;
-    Integer percent;
+    Double percent;
 }
